% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mrs.R
\docType{data}
\name{data_mrs}
\alias{data_mrs}
\title{MRS data}
\format{A data frame with 107 rows and 3 variables
\describe{
  \item{mass}{the mass of animals}
  \item{speed}{the speed of animals}
  \item{hopper}{the indicator variabel of hoppers}
}}
\source{
Garland, T.(1983).
}
\usage{
data(data_mrs)
}
\description{
A dataset containing the mass, speed and hopper indicator for land animals.
 The variables are as follows:
}
\examples{
\dontrun{
data(data_mrs)
summary(data_mrs)
}
}
\references{
\itemize{
\item Garland, T.(1983).
  The relation between maximal running speed and body
  mass in terrestrial mammals. Journal of Zoology 199, 157--170.
  }
}
\keyword{data_mrs}

